requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pypi_0
axial-positional-embedding=0.2.1=pypi_0
backcall=0.2.0=pypi_0
bertviz=1.3.0=pypi_0
black=22.1.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blessings=1.7=pypi_0
boto3=1.21.25=pypi_0
botocore=1.24.25=pypi_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dgl-cuda11.1=0.8.1=py38_0
dill=0.3.4=pypi_0
dm-tree=0.1.6=pypi_0
einops=0.4.0=pypi_0
entrypoints=0.4=pypi_0
etils=0.4.0=pypi_0
executing=0.8.2=pypi_0
filelock=3.6.0=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.30.0=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20220512=pypi_0
gast=0.3.3=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.56.0=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=2.10.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.1=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
ipykernel=6.9.1=pypi_0
ipython=8.0.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
jmespath=1.0.0=pypi_0
joblib=1.1.0=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=13.0.0=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
local-attention=1.4.3=pypi_0
lz4=4.0.0=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.1.0=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.11=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pypi_0
networkx=2.7.1=pyhd3eb1b0_0
notebook=6.4.8=pypi_0
numpy=1.22.3=pypi_0
numpy-base=1.21.5=py38hf524024_2
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1o=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
performer-pytorch=1.1.4=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=py38h06a4308_0
platformdirs=2.5.1=pypi_0
portalocker=2.4.0=pypi_0
product-key-memory=0.1.10=pypi_0
prometheus-client=0.13.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.28=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.11.2=pypi_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1.2=h7f8727e_1
reformer-pytorch=1.4.4=pypi_0
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
s3transfer=0.5.2=pypi_0
sacremoses=0.0.47=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py38h06a4308_0
sh=1.14.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
stack-data=0.2.0=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-datasets=4.5.2=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
tensorflow-metadata=1.7.0=pypi_0
tensorflow-text=2.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
thop=0.0.31-2005241907=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.8.1rc2=pypi_0
torch=1.10.0+cu111=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.13=pypi_0
torchaudio=0.10.0+rocm4.1=pypi_0
torchvision=0.11.0+cu111=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
transformers=3.3.1=pypi_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.8=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wikiextractor=3.0.6=pypi_0
wrapt=1.13.3=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
